Create Folder + Readme.md file (WhatsApp)
Good night. I include a step-by-step description of the functionality for integration with WhatsApp. Thanks to Javier Cristancho, who told me how to do it. 
Please, I appreciate if you can help me to adjust the folder numbers, I did not find how to do it. Thank you so much!